+++ src/SimpleTestSuiteFactory.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace MyTester;
 
+use MyTester\PHPT\PHPTTestCase;
+
 /**
  * Default test suite factory for {@see Tester}
  *
@@ -12,10 +14,13 @@ namespace MyTester;
  */
 final class SimpleTestSuiteFactory implements ITestSuiteFactory
 {
-    public function create(string $className): TestCase
+    public function create(string $className): ?TestCase
     {
         if (!is_subclass_of($className, TestCase::class)) {
             throw new InvalidTestSuiteException("$className is not a descendant of " . TestCase::class . ".");
+        }
+        if ($className === PHPTTestCase::class) {
+            return null;
         }
         return new $className();
     }